backend/requirements.txt
flask
flask-cors
requests
requests-cache
python-dotenv
pytest
requests-mock
